[#9279] YSQL: Enable -Wextra on pggate
Summary:
As a part of enabling `-Wextra` and some other flags from RocksDB on entire code base this diff enables flags on pggate directory.

All new warnings are fixed in context of this diff. Compilation was tested on GCC9.

Test Plan: Jenkins

Reviewers: mbautin, timur, sergei

Reviewed By: sergei

Subscribers: sergei, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D12341

+++ src/yb/util/uuid.h
@@ -143,7 +143,7 @@ class Uuid {
     }
     if (boost_uuid_.version() == boost::uuids::uuid::version_time_based) {
       // Compare the hi timestamp bits.
-      for (int i = 6; i < kUuidMsbSize; i++) {
+      for (size_t i = 6; i < kUuidMsbSize; i++) {
         if (boost_uuid_.data[i] < other.boost_uuid_.data[i]) {
           return true;
         } else if (boost_uuid_.data[i] > other.boost_uuid_.data[i]) {
@@ -168,7 +168,7 @@ class Uuid {
       }
     } else {
       // Compare all the other bits
-      for (int i = 0; i < kUuidMsbSize; i++) {
+      for (size_t i = 0; i < kUuidMsbSize; i++) {
         if (boost_uuid_.data[i] < other.boost_uuid_.data[i]) {
           return true;
         } else if (boost_uuid_.data[i] > other.boost_uuid_.data[i]) {
@@ -178,7 +178,7 @@ class Uuid {
     }
 
     // Then compare the remaining bytes.
-    for (int i = kUuidMsbSize; i < kUuidSize; i++) {
+    for (size_t i = kUuidMsbSize; i < kUuidSize; i++) {
       if (boost_uuid_.data[i] < other.boost_uuid_.data[i]) {
         return true;
       } else if (boost_uuid_.data[i] > other.boost_uuid_.data[i]) {
